+++ packages/ricos-editor/web/src/RicosContext/locale-resource-provider.tsx
@@ -1,19 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import type { TranslationFunction } from 'ricos-types';
 import { RicosTranslate } from 'wix-rich-content-common';
-import { RicosContextProvider } from 'ricos-context';
 import { fetchLocaleResource } from './translations';
 
-export const LocaleResourceProvider = ({
-  children,
-  isMobile,
-  locale,
-  localeContent,
-  experiments,
-  languageDir,
-  theme,
-  portal,
-}) => {
+export default function LocaleResourceProvider({ locale, children }) {
   const [currentLocaleResource, setLocaleResource] = useState<{
     locale: string;
     localeResource?: Record<string, string>;
@@ -26,14 +15,14 @@ export const LocaleResourceProvider = ({
 
   // custom hook
   useEffect(() => {
-    const hanldeLocaleResource = async () => {
+    const handleLocaleResource = async () => {
       const localeResource = await fetchLocaleResource(locale);
       if (localeResource) {
         setLocaleResource(localeResource);
       }
     };
 
-    hanldeLocaleResource();
+    handleLocaleResource();
   }, []);
 
   if (!currentLocaleResource) {
@@ -44,20 +33,7 @@ export const LocaleResourceProvider = ({
       locale={currentLocaleResource.locale}
       localeResource={currentLocaleResource.localeResource}
     >
-      {(t: TranslationFunction) => (
-        <RicosContextProvider
-          t={t}
-          locale={currentLocaleResource.locale}
-          localeContent={localeContent}
-          isMobile={isMobile}
-          experiments={experiments}
-          languageDir={languageDir}
-          theme={theme}
-          portal={portal}
-        >
-          {children}
-        </RicosContextProvider>
-      )}
+      {children(currentLocaleResource.locale)}
     </RicosTranslate>
   );
-};
+}